# wxrc: Faster&leaner compilation of C++ source file generated by resource compiler.