@vadz 2024-04-08
It looks like this was superseded by #24437, so closing.